+++ frontend/src/layouts/DefaultLayout.vue
@@ -5,6 +5,8 @@
 import { computed } from 'vue'
 import { useRouter } from 'vue-router'
 
+import { ArrowDown, Moon, Sunny } from '@element-plus/icons-vue'
+
 import { useLanguage } from '@/composables'
 import { toolbarShortForUiCode } from '@/i18n/locales'
 import { useAuthStore, useUIStore } from '@/stores'